♻️ refactor: 코드 스타일 규칙에 맟게 atom 경로 이동
