+++ src/renderer/src/features/champ/ChampDetail/ChampDetail.tsx
@@ -16,7 +16,7 @@ import TierIcon, { HoneyIcon, OpIcon } from '@renderer/features/asset/TierIcon'
 import ItemBuildGroups from '@renderer/features/item/ItemBuildGroups'
 import LaneSelect from '@renderer/features/lane/LaneSelect'
 import RankRangeSelect from '@renderer/features/rank/RankRangeSelect'
-import { rankRangeIdAtom } from '@renderer/features/rank/RankRangeSelect/rankRangeId.atom'
+import { rankRangeIdAtom } from '@renderer/features/rank/RankRangeSelect/atoms/rankRangeId.atom'
 import RuneBuildButtonRadioList from '@renderer/features/rune/RuneBuildButtonRadioList'
 import RunePage from '@renderer/features/rune/RunePage'
 import useAPI from '@renderer/hooks/useAPI'
