+++ src/renderer/src/features/duo/DuoSynergyTable/DuoSynergyTable.tsx
@@ -19,10 +19,10 @@ import {
 import ChampionProfileSmall from '@renderer/features/asset/ChampionProfileSmall'
 import LaneIcon from '@renderer/features/asset/LaneIcon/LaneIcon'
 import DuoLaneSelect, { DUO_OPTIONS, DuoId } from '@renderer/features/duo/DuoLaneSelect'
-import { duoSynergyTableDuoIdAtom } from '@renderer/features/duo/DuoSynergyTable/duoSynergyTableDuoId.atom'
+import { duoSynergyTableDuoIdAtom } from '@renderer/features/duo/DuoSynergyTable/atoms/duoSynergyTableDuoId.atom'
 import { LANE_LABELS } from '@renderer/features/lane/LaneSelect'
 import RankRangeSelect from '@renderer/features/rank/RankRangeSelect'
-import { rankRangeIdAtom } from '@renderer/features/rank/RankRangeSelect/rankRangeId.atom'
+import { rankRangeIdAtom } from '@renderer/features/rank/RankRangeSelect/atoms/rankRangeId.atom'
 import useAPI from '@renderer/hooks/useAPI'
 import useCustomForm from '@renderer/hooks/useCustomForm'
 import useDidUpdateEffect from '@renderer/hooks/useDidUpdateEffect'
