+++ src/renderer/src/features/tier/TierTable/TierTable.tsx
@@ -13,8 +13,8 @@ import RankingVariation from '@renderer/components/RankingVariation'
 import ChampionProfileSmall from '@renderer/features/asset/ChampionProfileSmall'
 import LaneSelect from '@renderer/features/lane/LaneSelect'
 import RankRangeSelect from '@renderer/features/rank/RankRangeSelect'
-import { rankRangeIdAtom } from '@renderer/features/rank/RankRangeSelect/rankRangeId.atom'
-import { tierTableLaneIdAtom } from '@renderer/features/tier/TierTable/tierTableLaneId.atom'
+import { rankRangeIdAtom } from '@renderer/features/rank/RankRangeSelect/atoms/rankRangeId.atom'
+import { tierTableLaneIdAtom } from '@renderer/features/tier/TierTable/atoms/tierTableLaneId.atom'
 import useAPI from '@renderer/hooks/useAPI'
 import useCustomForm from '@renderer/hooks/useCustomForm'
 import useDidUpdateEffect from '@renderer/hooks/useDidUpdateEffect'
